LF-4195 separate creatable select and select into different components
